Set up basic app skeleton
This includes setting up a 'TheContainer' component which will serve as the place for the app's content.
We've also included a Suspense component within here, to display a, very basic for now, loading screen while
child components are setting up.
Lastly, we've set up basic routing that will redirect from '/' to '/invoices' and load in a very basic
ViewInvoiceList component that will for now simulate an API call and render just the number for some dummy invoices.